backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
python-multipart==0.0.6
cors==1.0.1
pydantic==2.5.0
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
sqlalchemy==2.0.23
alembic==1.13.0  # For database migrations (optional)